Relocate 'AstLineCounter'
